move bb issue link to header, simplify templating
the old template was creating a header and a footer.

this one is only creating a header (so less boilerplate) and just
links the words "Original report" to the bitbucket url.

the attachments list (if any) is also in the header.

additionally to that, I made the templating easier by using a full template
instead of splitting into header / content / footer - so stuff is easier to
modify.

without bb_skip, it would have been even simpler, but i did not remove that.
to support it, i added a second full template for report and comment.

the change is practically tested, i used this code to migrate to:

https://github.com/moinwiki/moin