Use Bitbucket API 2.0
Per https://developer.atlassian.com/cloud/bitbucket/deprecation-notice-v1-apis/,
Bitbucket API version 1.0 is deprecated as of June 30 2018
and will no longer work in April of 2019.  Alter the issue
fetch methods to use the v2.0 API.

Additionally, the "issues" collection is now streamed through
the tool in order to support large issue sets with thousands
of issues without a large memory footprint.   Added a new
option --mention-changes that formats status changes as comments;
the new Bitbucket API now allows us to have these status changes
available.

I did notice that the "/changes/" API seems to throw a 500
on at least one issue persistently.   The issue in question
seems to have a broken assignee on it, so since this is a new
API for BB that still has some bugs, it warns for 500
on just the "changes" API and skips.