Add option to upload attachments to wiki
Since github doesn't allow attachments to be uploaded through
the API, and requests for this to be added have been ignored for
years, work around it by pushing attachment files to the
wiki git repo.    Add links using relative URLs into the
issue so they are both linked and also portable with the
repo.